# Support several sequence representations