--- a/src/boot/lib/ast.ml
+++ b/src/boot/lib/ast.ml
@@ -113,6 +113,11 @@ and const =
   | Ctail
   | Cnull
   | Cmap of (tm -> tm) option
+  | Cmapi of (int -> tm -> tm) option
+  | Citer of (tm -> unit) option
+  | Citeri of (int -> tm -> unit) option
+  | Cfoldl of (tm -> tm -> tm) option * tm option
+  | Cfoldr of (tm -> tm -> tm) option * tm option
   | Csubsequence of tm Mseq.t option * int option
   (* MCore intrinsics: Random numbers *)
   | CrandIntU of int option
@@ -551,6 +556,11 @@ let const_has_side_effect = function
   | Ctail
   | Cnull
   | Cmap _
+  | Cmapi _
+  | Citer _
+  | Citeri _
+  | Cfoldl _
+  | Cfoldr _
   | Csubsequence _ ->
       false
   (* MCore intrinsics: Random numbers *)