--- a/src/boot/lib/bootparser.ml
+++ b/src/boot/lib/bootparser.ml
@@ -122,22 +122,28 @@ let symbolizeEnvWithKeywords keywords =
 
 let parseMExprString keywords str =
   PTreeTm
-    ( str |> Parserutils.parse_mexpr_string
+    ( str |> Intrinsics.Mseq.Helpers.to_ustring
+    |> Parserutils.parse_mexpr_string
     |> Parserutils.raise_parse_error_on_non_unique_external_id
     |> Symbolize.symbolize
          (builtin_name2sym @ symbolizeEnvWithKeywords keywords)
     |> Parserutils.raise_parse_error_on_partially_applied_external )
 
 let parseMCoreFile keywords filename =
+  let filename = filename in
   let symKeywordsMap = symbolizeEnvWithKeywords keywords in
   let name2sym = builtin_name2sym @ symKeywordsMap in
   let symKeywords = List.map (fun (_, s) -> s) symKeywordsMap in
-  PTreeTm
-    ( filename |> Parserutils.parse_mcore_file
-    |> Parserutils.raise_parse_error_on_non_unique_external_id
-    |> Symbolize.symbolize name2sym
-    |> Deadcode.elimination builtin_sym2term name2sym symKeywords
-    |> Parserutils.raise_parse_error_on_partially_applied_external )
+  let res =
+    PTreeTm
+      ( filename |> Intrinsics.Mseq.Helpers.to_ustring
+      |> Parserutils.parse_mcore_file
+      |> Parserutils.raise_parse_error_on_non_unique_external_id
+      |> Symbolize.symbolize name2sym
+      |> Deadcode.elimination builtin_sym2term name2sym symKeywords
+      |> Parserutils.raise_parse_error_on_partially_applied_external )
+  in
+  res
 
 (* Returns a tuple with the following elements
    1. ID field