--- a/src/boot/lib/builtin.ml
+++ b/src/boot/lib/builtin.ml
@@ -73,6 +73,7 @@ let builtin =
   ; ("snoc", f (Csnoc None))
   ; ("splitAt", f (CsplitAt None))
   ; ("reverse", f Creverse)
+  ; ("tail", f Ctail)
   ; ("subsequence", f (Csubsequence (None, None)))
     (* MCore intrinsics: Random numbers *)
   ; ("randIntU", f (CrandIntU None))