--- a/src/boot/lib/intrinsics.ml
+++ b/src/boot/lib/intrinsics.ml
@@ -19,17 +19,17 @@ module Mseq = struct
 
   let empty_fingertree = FingerTree BatFingerTree.empty
 
-  let create = create_fingertree
-
-  let empty = empty_fingertree
+  (* let create = create_fingertree
+   *
+   * let empty = empty_fingertree *)
 
   (* let create = create_list
-   * 
+   *
    * let empty = empty_list *)
 
-  (* let create = create_rope
-   *
-   * let empty = empty_rope *)
+  let create = create_rope
+
+  let empty = empty_rope
 
   let length = function
     | Rope s ->
@@ -288,23 +288,23 @@ module Mseq = struct
       | _ ->
           raise (Invalid_argument "Mseq.fold_right2")
 
-    let of_list = of_list_fingertree
-
-    let of_array = of_array_fingertree
-
-    let of_ustring = of_ustring_fingertree
+    (* let of_list = of_list_fingertree
+     *
+     * let of_array = of_array_fingertree
+     *
+     * let of_ustring = of_ustring_fingertree *)
 
     (* let of_list = of_list_list
-     * 
+     *
      * let of_array = of_array_list
-     * 
+     *
      * let of_ustring = of_ustring_list *)
 
-    (* let of_list = of_list_rope
-     *
-     * let of_array = of_array_rope
-     *
-     * let of_ustring = of_ustring_rope *)
+    let of_list = of_list_rope
+
+    let of_array = of_array_rope
+
+    let of_ustring = of_ustring_rope
   end
 end