--- a/src/boot/lib/intrinsics.mli
+++ b/src/boot/lib/intrinsics.mli
@@ -2,13 +2,21 @@ open Ustring.Op
 
 module Mseq : sig
   type 'a t =
-    | FingerTreeSeq of 'a BatFingerTree.t
-    | ListSeq of 'a List.t
-    | RopeSeq of 'a array Rope.t
+    | FingerTree of 'a BatFingerTree.t
+    | List of 'a List.t
+    | Rope of 'a array Rope.t
 
   val create : int -> (int -> 'a) -> 'a t
 
+  val create_list : int -> (int -> 'a) -> 'a t
+
+  val create_rope : int -> (int -> 'a) -> 'a t
+
   val empty : 'a t
+
+  val empty_rope : 'a t
+
+  val empty_list : 'a t
 
   val length : 'a t -> int
 
@@ -31,13 +39,25 @@ module Mseq : sig
   module Helpers : sig
     val of_list : 'a list -> 'a t
 
+    val of_list_rope : 'a list -> 'a t
+
+    val of_list_list : 'a list -> 'a t
+
     val to_list : 'a t -> 'a list
 
     val of_array : 'a array -> 'a t
 
+    val of_array_rope : 'a array -> 'a t
+
+    val of_array_list : 'a array -> 'a t
+
     val to_array : 'a t -> 'a array
 
     val of_ustring : ustring -> int t
+
+    val of_ustring_rope : ustring -> int t
+
+    val of_ustring_list : ustring -> int t
 
     val to_ustring : int t -> ustring