* Adds a finger tree and cons list sequence representation. A tag keeps track of whether a representation is a rope, finger tree or cons list. Runtime errors are given if one tries to use two sequences of different representations together (e.g. `concat`).
* Intrinsic functions `createFingerTree`, `createList` and `createRope` are added, to create the respective sequence version. `create` is synonymous with `createRope` for now.
* Two classes of microbenchmarks are added, for comparing `map` and `fold` between OCaml and MCore for the three sequence versions.
* I experimented with adding `tail`, `null`, `head`, and `map` as intrinsic functions instead of implementing them in `seq.mc`. See the result on the micro benchmark in a comment below. I haven't implemented evaluation rules in `eval.mc` for these intrinsics yet but I'll do that for those that we decide to keep, if any.
* Small bugfix in `bootparser.ml`.
* I replaced `makeSeq` with `create` in `boot-parser.mc` and fixed indentation in that file.
* Adds a `mergeSort` implementation to `seq.mc`