@lingmar 2021-05-28
Benchmarking results for bootstrapping

**With tagging (this PR commit cf982c7)**
```
for i in {1..10}; do time mi compile src/main/mi.mc; done
mi compile src/main/mi.mc  14.32s user 0.59s system 99% cpu 14.985 total
mi compile src/main/mi.mc  13.90s user 0.58s system 99% cpu 14.523 total
mi compile src/main/mi.mc  13.92s user 0.57s system 99% cpu 14.507 total
mi compile src/main/mi.mc  14.27s user 0.57s system 99% cpu 14.900 total
mi compile src/main/mi.mc  14.32s user 0.60s system 99% cpu 14.970 total
mi compile src/main/mi.mc  14.26s user 0.59s system 99% cpu 14.933 total
mi compile src/main/mi.mc  14.54s user 0.61s system 99% cpu 15.202 total
mi compile src/main/mi.mc  13.90s user 0.58s system 99% cpu 14.554 total
mi compile src/main/mi.mc  13.92s user 0.58s system 99% cpu 14.516 total
mi compile src/main/mi.mc  13.87s user 0.57s system 99% cpu 14.477 total
```
sum: 147.567 s

**Without tagging ([this branch](https://github.com/lingmar/miking/tree/no-tagging) commit 426741f)**
```
for i in {1..10}; do time mi compile src/main/mi.mc; done
mi compile src/main/mi.mc  15.09s user 0.60s system 99% cpu 15.711 total
mi compile src/main/mi.mc  14.38s user 0.60s system 99% cpu 15.015 total
mi compile src/main/mi.mc  14.22s user 0.57s system 99% cpu 14.824 total
mi compile src/main/mi.mc  14.27s user 0.59s system 99% cpu 14.925 total
mi compile src/main/mi.mc  14.08s user 0.56s system 99% cpu 14.695 total
mi compile src/main/mi.mc  14.05s user 0.57s system 99% cpu 14.659 total
mi compile src/main/mi.mc  13.94s user 0.57s system 99% cpu 14.541 total
mi compile src/main/mi.mc  14.02s user 0.56s system 99% cpu 14.621 total
mi compile src/main/mi.mc  14.13s user 0.58s system 99% cpu 14.740 total
mi compile src/main/mi.mc  14.61s user 0.61s system 99% cpu 15.306 total
```
sum: 149.037 s