@lingmar 2021-06-01
Results for `map` and `fold` micro benchmarks. List lengths of 10000 were used.

Summary:
* `tail`: makes a huge difference for List and FingerTree.
* `null`: important for List.
* `head`: no noticable difference.
* `map`: huge difference for `Rope`, some difference for `FingerTree`

**Fingertree**
_024c769 Add map and fold micro benchmarks_
```
./run map_fingertree 5 12
3. Miking compiler:     26.902081s 25.939930s
4. Ocaml byte code       0.037225s  0.021171s
5. OCaml native:         0.098911s  0.007807s

./run fold_fingertree 5 12
3. Miking compiler:     18.472092s 18.012835s
4. Ocaml byte code       0.017493s  0.016604s
5. OCaml native:         0.195401s  0.006362s
```

_a6a275b Add tail as an intrinsic function_
```
./run map_fingertree 5 12
3. Miking compiler:      0.285461s  0.022976s
4. Ocaml byte code       0.016444s  0.016072s
5. OCaml native:         0.203734s  0.006374s

./run fold_fingertree 5 12
3. Miking compiler:      0.187129s  0.014041s
4. Ocaml byte code       0.017504s  0.016104s
5. OCaml native:         0.096347s  0.006977s
```

_6a726fd Add null as intrinsic function_

```
./run map_fingertree 5 12
3. Miking compiler:      0.114255s  0.024146s
4. Ocaml byte code       0.017217s  0.016395s
5. OCaml native:         0.093580s  0.007138s

./run fold_fingertree 5 12
3. Miking compiler:      0.103591s  0.014374s
4. Ocaml byte code       0.018660s  0.023775s
5. OCaml native:         0.093169s  0.006419s
```

_3ed30bb Add head as intrinsic function_

```
./run map_fingertree 5 12
3. Miking compiler:      0.106938s  0.019657s
4. Ocaml byte code       0.016330s  0.015996s
5. OCaml native:         0.096175s  0.006459s

./run fold_fingertree 5 12
3. Miking compiler:      0.101025s  0.014723s
4. Ocaml byte code       0.016762s  0.016756s
5. OCaml native:         0.098508s  0.009405s
```

_9fb8db4 Add map as intrinsic function_

```
./run map_fingertree 5 12
3. Miking compiler:      0.099310s  0.014251s
4. Ocaml byte code       0.016765s  0.016635s
5. OCaml native:         0.091480s  0.007175s

./run fold_fingertree 5 12
3. Miking compiler:      0.104867s  0.014021s
4. Ocaml byte code       0.017141s  0.016180s
5. OCaml native:         0.101865s  0.011247s
```

**List**
_024c769 Add map and fold micro benchmarks_
```
./run map_list 5 12
3. Miking compiler:     14.583236s 14.140232s
4. Ocaml byte code       0.020532s  0.019372s
5. OCaml native:         0.086057s  0.008806s

./run fold_list 5 12
3. Miking compiler:      3.547316s  3.066350s
4. Ocaml byte code       0.014637s  0.014138s
5. OCaml native:         0.168814s  0.005984s
```

_a6a275b Add tail as an intrinsic function_
```
./run map_list 5 12
3. Miking compiler:      0.489501s  0.311097s
4. Ocaml byte code       0.020603s  0.020130s
5. OCaml native:         0.085282s  0.008408s

./run fold_list 5 12
3. Miking compiler:      0.493672s  0.302565s
4. Ocaml byte code       0.015274s  0.014683s
5. OCaml native:         0.091747s  0.008037s
```

_6a726fd Add null as intrinsic function_

```
./run map_list 5 12
3. Miking compiler:      0.100858s  0.014049s
4. Ocaml byte code       0.020808s  0.020871s
5. OCaml native:         0.086594s  0.008311s

./run fold_list 5 12
3. Miking compiler:      0.100019s  0.009935s
4. Ocaml byte code       0.015443s  0.015849s
5. OCaml native:         0.098575s  0.006642s
```

_3ed30bb Add head as intrinsic function_

```
./run map_list 5 12
3. Miking compiler:      0.106938s  0.019657s
4. Ocaml byte code       0.016330s  0.015996s
5. OCaml native:         0.096175s  0.006459s

./run fold_list 5 12
3. Miking compiler:      0.100897s  0.009011s
4. Ocaml byte code       0.015791s  0.016091s
5. OCaml native:         0.097764s  0.006809s
```

_9fb8db4 Add map as intrinsic function_

```
./run map_list 5 12
3. Miking compiler:      0.095731s  0.009508s
4. Ocaml byte code       0.019759s  0.019520s
5. OCaml native:         0.087915s  0.008348s

./run fold_list 5 12
3. Miking compiler:      0.104127s  0.010755s
4. Ocaml byte code       0.016716s  0.017297s
5. OCaml native:         0.101933s  0.008999s
```

**Rope**
_024c769 Add map and fold micro benchmarks_

```
./run map_rope 5 12
3. Miking compiler:      2.540888s  2.329109s
4. Ocaml byte code       0.014852s  0.014221s
5. OCaml native:         0.191961s  0.005958s

./run fold_rope 5 12
3. Miking compiler:      1.612828s  1.396937s
4. Ocaml byte code       0.014591s  0.014978s
5. OCaml native:         0.191444s  0.005983s
```

_a6a275b Add tail as an intrinsic function_
```
./run map_rope 5 12
3. Miking compiler:      2.559553s  2.325282s
4. Ocaml byte code       0.014918s  0.014517s
5. OCaml native:         0.085244s  0.006045

./run fold_rope 5 12`
3. Miking compiler:      1.764123s  1.491369s
4. Ocaml byte code       0.015100s  0.016090s
5. OCaml native:         0.089939s  0.006665s
```

_6a726fd Add null as intrinsic function_

```
./run map_rope 5 12
3. Miking compiler:      2.364776s  2.238750s
4. Ocaml byte code       0.022357s  0.014958s
5. OCaml native:         0.085419s  0.006166s

./run fold_rope 5 12
3. Miking compiler:      1.810326s  1.572013s
4. Ocaml byte code       0.015996s  0.022434s
5. OCaml native:         0.094882s  0.007247s
```

_3ed30bb Add head as intrinsic function_

```
./run map_rope 5 12
3. Miking compiler:      2.366129s  2.250198s
4. Ocaml byte code       0.014370s  0.018756s
5. OCaml native:         0.086600s  0.006009s

./run fold_rope 5 12
3. Miking compiler:      1.886065s  1.697783s
4. Ocaml byte code       0.015201s  0.022637s
5. OCaml native:         0.090819s  0.006770s
```

_9fb8db4 Add map as intrinsic function_

```
./run map_rope 5 12
3. Miking compiler:      0.094614s  0.007869s
4. Ocaml byte code       0.018605s  0.018594s
5. OCaml native:         0.084300s  0.005936s

./run fold_rope 5 12
3. Miking compiler:      1.892593s  1.709174s
4. Ocaml byte code       0.014968s  0.016785s
5. OCaml native:         0.089345s  0.006746s
```